CloudFunction2/packaging/requirements.txt
pboto3==1.6.19
botocore==1.9.23